Split up code into different files
